# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SuiMedrassa – School Management SaaS API
  version: 1.0.0
  description: |
    Multi-tenant school management platform.

    ## Authentication
    All protected endpoints require a **Bearer token** obtained from `POST /api/v1/auth/login`.
    Pass it as `Authorization: Bearer <token>`.

    ## Role legend
    | Role | Scope |
    |------|-------|
    | `system_admin` | Global – all schools |
    | `school_owner` | Own school |
    | `admin` | Own school (cannot delete school) |
    | `teacher` | Own school – read-only on most resources |
    | `student` | Own school – own data only |

    ## Tenant context
    For `system_admin` acting on a specific school, pass the `X-School-Id` header.
  contact:
    email: admin@suimedrassa.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.suimedrassa.com/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Sanctum personal access token

  # ── Reusable parameters ───────────────────────────────────────────────────
...
  schemas:
...
    User:
      type: object
      properties:
        id:            { type: integer }
        school_id:     { type: integer, nullable: true }
        role:          { $ref: '#/components/schemas/UserRole' }
        name:          { type: string }
        email:         { type: string, format: email }
        phone:         { type: string, nullable: true }
        avatar:        { type: string, nullable: true }
        is_active:     { type: boolean }
        last_login_at: { type: string, format: date-time, nullable: true }
        created_at:    { type: string, format: date-time }
        updated_at:    { type: string, format: date-time }
...
    AuthToken:
      type: object
      properties:
        user:  { $ref: '#/components/schemas/User' }
        token: { type: string, example: "1|abcdef..." }
...
security:
  - bearerAuth: []
paths:
...
  /auth/register-school:
    post:
      tags: [Auth]
      summary: Register a new school (public)
      description: |
        Creates a new school with a 14-day trial and a `school_owner` user in one atomic call.
        **No authentication required.**
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [school_name, school_email, owner_name, owner_email, password, password_confirmation]
              properties:
                school_name:           { type: string, example: "Springfield High" }
                school_email:          { type: string, format: email }
                owner_name:            { type: string, example: "John Smith" }
                owner_email:           { type: string, format: email }
                password:              { type: string, format: password, minLength: 8 }
                password_confirmation: { type: string, format: password }
                phone:                 { type: string, nullable: true }
                country:               { type: string, nullable: true }
      responses:
        "201":
          description: School & owner created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:   { $ref: '#/components/schemas/User' }
                  school: { $ref: '#/components/schemas/School' }
                  token:  { type: string }
        "422": { $ref: '#/components/responses/ValidationError' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Login (all roles, public)
      description: Returns a Sanctum bearer token. Previous tokens are invalidated.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string, format: password }
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthToken'
        "403":
          description: Account or school is inactive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "422": { $ref: '#/components/responses/ValidationError' }
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Send password reset email (public)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "200":
          description: Reset link sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        "422": { $ref: '#/components/responses/ValidationError' }
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password using email token (public)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, email, password, password_confirmation]
              properties:
                token:                 { type: string }
                email:                 { type: string, format: email }
                password:              { type: string, format: password, minLength: 8 }
                password_confirmation: { type: string, format: password }
      responses:
        "200":
          description: Password reset successfully
        "422": { $ref: '#/components/responses/ValidationError' }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout (revoke current token)
      description: "**Roles:** all authenticated"
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /auth/me:
    get:
      tags: [Profile]
      summary: Get authenticated user profile
      description: "**Roles:** all authenticated"
      responses:
        "200":
          description: User with school and roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "401": { $ref: '#/components/responses/Unauthorized' }
    patch:
      tags: [Profile]
      summary: Update own profile (name, phone, password)
      description: "**Roles:** all authenticated"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:             { type: string }
                phone:            { type: string, nullable: true }
                current_password: { type: string, format: password }
                password:         { type: string, format: password, minLength: 8 }
                password_confirmation: { type: string, format: password }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "422": { $ref: '#/components/responses/ValidationError' }

  # ════════════════════════════════════════════════════════════════════════════
  # SYSTEM ADMIN
  # ════════════════════════════════════════════════════════════════════════════
  /system/dashboard:
    get:
      tags: [System – Dashboard]
      summary: Platform-wide dashboard statistics
      description: "**Roles:** `system_admin`"
      responses:
        "200":
          description: Global statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemDashboard'
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /system/stats:
    get:
      tags: [System – Dashboard]
      summary: Quick global counts (legacy)
      description: "**Roles:** `system_admin`"
      responses:
        "200":
          description: Counts of schools and users
          content:
            application/json:
              schema:
                type: object
                properties:
                  schools:        { type: integer }
                  active_schools: { type: integer }
                  trial_schools:  { type: integer }
                  suspended:      { type: integer }
                  total_users:    { type: integer }
                  total_students: { type: integer }
                  total_teachers: { type: integer }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /system/schools:
    get:
      tags: [System – Schools]
      summary: List all schools
      description: "**Roles:** `system_admin`"
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended, trial, cancelled]
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Paginated schools
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/School' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /system/schools/{school}:
    parameters:
      - name: school
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [System – Schools]
      summary: Get school details
      description: "**Roles:** `system_admin`"
      responses:
        "200":
          description: School with plan
          content:
            application/json:
              schema: { $ref: '#/components/schemas/School' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [System – Schools]
      summary: Update school (status, tier, plan, dates)
      description: "**Roles:** `system_admin`"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:                 { type: string }
                email:                { type: string, format: email }
                phone:                { type: string, nullable: true }
                address:              { type: string, nullable: true }
                status:
                  type: string
                  enum: [active, suspended, trial, cancelled]
                subscription_tier:
                  type: string
                  enum: [tier1, tier2, tier3, trial]
                subscription_plan_id: { type: integer, nullable: true }
                trial_ends_at:        { type: string, format: date-time, nullable: true }
                subscription_ends_at: { type: string, format: date-time, nullable: true }
      responses:
        "200":
          description: Updated school
          content:
            application/json:
              schema: { $ref: '#/components/schemas/School' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "422": { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [System – Schools]
      summary: Soft-delete a school
      description: "**Roles:** `system_admin`"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /system/schools/{school}/subscription:
    parameters:
      - name: school
        in: path
        required: true
        schema: { type: integer }
    patch:
      tags: [System – Schools]
      summary: Change school subscription tier/plan
      description: "**Roles:** `system_admin`"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subscription_tier]
              properties:
                subscription_tier:
                  type: string
                  enum: [tier1, tier2, tier3, trial]
                subscription_plan_id:  { type: integer, nullable: true }
                subscription_ends_at:  { type: string, format: date-time, nullable: true }
      responses:
        "200":
          description: Updated school
          content:
            application/json:
              schema: { $ref: '#/components/schemas/School' }
        "422": { $ref: '#/components/responses/ValidationError' }
  /system/plans:
    get:
      tags: [System – Subscription Plans]
      summary: List all subscription plans
      description: "**Roles:** `system_admin`"
      responses:
        "200":
          description: List of plans
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SubscriptionPlan' }
    post:
      tags: [System – Subscription Plans]
      summary: Create a subscription plan
      description: "**Roles:** `system_admin`"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug, price_monthly, price_yearly]
              properties:
                name:             { type: string }
                slug:             { type: string }
                description:      { type: string, nullable: true }
                max_students:     { type: integer, nullable: true }
                max_teachers:     { type: integer, nullable: true }
                max_classes:      { type: integer, nullable: true }
                storage_limit_mb: { type: integer }
                price_monthly:    { type: number }
                price_yearly:     { type: number }
                is_active:        { type: boolean }
      responses:
        "201":
          description: Plan created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubscriptionPlan' }
        "422": { $ref: '#/components/responses/ValidationError' }
  /system/plans/{subscriptionPlan}:
    parameters:
      - name: subscriptionPlan
        in: path
        required: true
        schema: { type: integer }
    patch:
      tags: [System – Subscription Plans]
      summary: Update a subscription plan
      description: "**Roles:** `system_admin`"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:             { type: string }
                description:      { type: string, nullable: true }
                max_students:     { type: integer, nullable: true }
                max_teachers:     { type: integer, nullable: true }
                max_classes:      { type: integer, nullable: true }
                storage_limit_mb: { type: integer }
                price_monthly:    { type: number }
                price_yearly:     { type: number }
                is_active:        { type: boolean }
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubscriptionPlan' }
        "422": { $ref: '#/components/responses/ValidationError' }
...
  /school/dashboard:
    get:
      tags: [School – Dashboard]
      summary: School-level statistics
      description: "**Roles:** `school_owner`, `admin`"
      parameters:
        - $ref: '#/components/parameters/XSchoolId'
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SchoolDashboard' }
...
  /school/students/{student}/classes:
    parameters:
      - name: student
        in: path
        required: true
        schema: { type: integer }
    get:
      tags: [School – Students]
      summary: Get classrooms a student is enrolled in
      description: "**Roles:** `school_owner`, `admin`"
      responses:
        "200":
          description: List of classrooms
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Classroom' }
...